Bump package version

Calls the following functions:

Verify that the current branch is the main branch.

update_news()

update_version(), using the which argument

Depending on the which argument:

If "dev", finalize_version() with push = FALSE

Otherwise, commit_version().

bump_version(which = "dev")

Arguments

which

Component of the version number to update. Supported values are "dev" (default), "patch", "minor" and "major".

Bumped too soon?

Have you just run bump_version(), then realized "oh shoot, I forgot to merge that PR"? Fear not, run unbump_version(), merge that PR, run bump_version().

See also

Examples

# Create temporary directory to hold a mock package.
tempdir <- file.path(tempdir(), "fledge", "bump")
dir.create(tempdir, recursive = TRUE)
# Create said mock package.
# Set open to TRUE if you want to play in the mock package.
pkg <- create_demo_project(open = FALSE, news = TRUE, dir = tempdir)
# Use functions as if inside the newly created package project.
# (Or go and actually run code inside the newly created package project!)
usethis::with_project(pkg, {
# Add a new R file.
usethis::use_r("cool-function", open = FALSE)
# Pretend we added useful code inside it.
# Track the new R file with Git.
gert::git_add("R/cool-function.R")
gert::git_commit("- Add cool function.")
# Bump version with fledge.
fledge::bump_version()
})
#> ✔ Setting active project to '/tmp/RtmpOggLGH/fledge/bump/tea'
#> • Edit 'R/cool-function.R'
#> • Call `use_test()` to create a matching test file
#> → Scraping 3 commit messages.
#> ✔ Found 1 NEWS-worthy entries.
#> 
#> ── Updating NEWS ──
#> 
#> → Adding new entries to NEWS.md.
#> Warning: 'Date' must be an ISO date: yyyy-mm-dd, but it is actually better to leave       this field out completely. It is not required.
#> 
#> ── Update Version ──
#> 
#> ✔ Package version bumped to 0.0.0.9001.
#> → Adding header to NEWS.md.
#> → Committing changes.
#> 
#> ── Tagging Version ──
#> 
#> → Creating tag v0.0.0.9001 with tag message derived from NEWS.md.
#> • Edit 'NEWS.md'
#> ! Call `fledge::finalize_version()`.
#> ✔ Setting active project to '<no active project>'
#> NULL
# Clean
unlink(tempdir, recursive = TRUE)