Automating CRAN release

init_release() and pre_release() are run when a milestone in the development of a package is reached and it is ready to be sent to CRAN.

release() sends to CRAN after performing several checks, and offers help with accepting the submission.

post_release() should be called after the submission has been accepted.

Usage

init_release(which = "next", force = FALSE)

pre_release(force = FALSE)

release()

post_release()

Arguments

which

Component of the version number to update. Supported values are

"next" ("major" if the current version is x.99.99.9yz, "minor" if the current version is x.y.99.za, "patch" otherwise),

"patch"

"minor",

"major".

force

Create branches and tags even if they exist. Useful to recover from a previously failed attempt.

Details

init_release():

Ensures that no modified files are in the git index.

Creates a release branch and bumps version to a non-development version which should be sent to CRAN.

Writes/updates cran-comments.md with useful information about the current release process.

Prompts the user to run urlchecker::url_update(), devtools::check_win_devel(), and rhub::check_for_cran().

pre_release():

Opens a pull request for the release branch for final checks.